completed delete confirm

+++ src/app/todo/todo.component.ts
@@ -98,6 +98,8 @@ export class TodoComponent implements OnInit {
   }
 
   deleteTodo(todo: Todo): void {
+    if (confirm("Are you sure you want to delete?")) {
+      // delete call goes here
       this.todoService.deleteTodo(todo)
         .subscribe(
         data => {
@@ -110,6 +112,7 @@ export class TodoComponent implements OnInit {
           this.errorMessage = <any>error;
           console.log('complete error', this.errorMessage);
         });
+    }
   }
 
   calculateOpenItems(): void {
